Issue #11079: Stricter code standard check
Motivation:

I have seen that people are loading Drupal services statically, which is
not a best practice. And the `--warning-severity` flag suppresses that
warning.

This commit will make sure that the developer has taken care of all
warnings in the files he has worked on.. so that one day we can remove
the `--warning-severity` flag from the CI check :)